Small bug fix
Made it impossible for the player to move the snake directly into it-
self (if the Snake's head is right next to another piece of the snake,
you can't deliberately make the head run into that piece of the snake,
unless it's already heading that way.
Example (x = snake, h = snake head, o = free space):
OOOOO
OXXOO
OXXOO
OHXOO
OOXOO
In this scenario, the player isn't able to make the snake go to the
right, as the snake would bite itself.

This was done to prevent players from accidentally eating themselves by
changing the direction twice within one tick, resulting in only the
last direction change being applied.